misc: rename `mystery-encounters` file (#7530)
