+++ test/tests/mystery-encounter/encounters/delibirdy-encounter.test.ts
@@ -19,8 +19,8 @@ import {
 import { DelibirdyEncounter } from "#mystery-encounters/delibirdy-encounter";
 import * as EncounterPhaseUtils from "#mystery-encounters/encounter-phase-utils";
 import { generateModifierType } from "#mystery-encounters/encounter-phase-utils";
+import * as MysteryEncounters from "#mystery-encounters/mystery-encounter-biomes";
 import type { MoneyRequirement } from "#mystery-encounters/mystery-encounter-requirements";
-import * as MysteryEncounters from "#mystery-encounters/mystery-encounters";
 import { MysteryEncounterPhase } from "#phases/mystery-encounter-phases";
 import { GameManager } from "#test/framework/game-manager";
 import { runMysteryEncounterToEnd, runSelectMysteryEncounterOption } from "#test/utils/encounter-test-utils";
