+++ test/tests/mystery-encounter/encounters/global-trade-system-encounter.test.ts
@@ -11,8 +11,8 @@ import { PokemonNatureWeightModifier } from "#modifiers/modifier";
 import * as EncounterPhaseUtils from "#mystery-encounters/encounter-phase-utils";
 import { generateModifierType } from "#mystery-encounters/encounter-phase-utils";
 import { GlobalTradeSystemEncounter } from "#mystery-encounters/global-trade-system-encounter";
-import * as MysteryEncounters from "#mystery-encounters/mystery-encounters";
-import { CIVILIZATION_ENCOUNTER_BIOMES } from "#mystery-encounters/mystery-encounters";
+import * as MysteryEncounters from "#mystery-encounters/mystery-encounter-biomes";
+import { CIVILIZATION_ENCOUNTER_BIOMES } from "#mystery-encounters/mystery-encounter-biomes";
 import { GameManager } from "#test/framework/game-manager";
 import { runMysteryEncounterToEnd } from "#test/utils/encounter-test-utils";
 import { ModifierSelectUiHandler } from "#ui/modifier-select-ui-handler";
